Experimental conversion fix

+++ hlsdk/dlls/vector.h
@@ -31,14 +31,14 @@ public:
         inline Vector2D operator*(float fl)                             const   { return Vector2D(x*fl, y*fl);  }
         inline Vector2D operator/(float fl)                             const   { fl=1/fl; return Vector2D(x*fl, y*fl);  }
         
-        inline float Length(void)                                               const   { return (float)sqrt(x*x + y*y );              }
+        inline float Length(void)                                               const   { return std::sqrt(x*x + y*y );              }
 
         inline Vector2D Normalize ( void ) const
         {
                 // Vector2D vec2;
 
                 float flLen = Length();
-                if ( flLen == 0 )
+                if ( flLen == 0.0f )
                 {
                         return Vector2D( 0, 0 );
                 }
@@ -62,13 +62,13 @@ class Vector                                            // same data-layout as e
 {                                                               //              which is a vec_t[3]
 public:
         // Construction/destruction
-        inline Vector(void): x(0.0), y(0.0), z(0.0)                                     { }
-        inline Vector(float X, float Y, float Z): x(0.0), y(0.0), z(0.0)        { x = X; y = Y; z = Z;                          }
+        inline Vector(void): x(0.0f), y(0.0f), z(0.0f)                                     { }
+        inline Vector(float X, float Y, float Z): x(0.0f), y(0.0f), z(0.0f)        { x = X; y = Y; z = Z;                          }
         //inline Vector(double X, double Y, double Z)                   { x = (float)X; y = (float)Y; z = (float)Z;     }
         //inline Vector(int X, int Y, int Z)                            { x = (float)X; y = (float)Y; z = (float)Z;     }
-        inline Vector(const Vector& v): x(0.0), y(0.0), z(0.0)  { x = v.x; y = v.y; z = v.z;                            } 
-        inline Vector(float rgfl[3]): x(0.0), y(0.0), z(0.0)    { x = rgfl[0]; y = rgfl[1]; z = rgfl[2];        }
-        inline Vector(const Vector2D& v): x(0.0), y(0.0), z(0.0){ x = v.x; y = v.y; }
+        inline Vector(const Vector& v): x(0.0f), y(0.0f), z(0.0f)  { x = v.x; y = v.y; z = v.z;                            } 
+        inline Vector(float rgfl[3]): x(0.0f), y(0.0f), z(0.0f)    { x = rgfl[0]; y = rgfl[1]; z = rgfl[2];        }
+        inline Vector(const Vector2D& v): x(0.0f), y(0.0f), z(0.0f){ x = v.x; y = v.y; }
 
         // Operators
         inline Vector operator-(void) const                     { return Vector(-x,-y,-z);                              }
@@ -81,7 +81,7 @@ public:
         
         // Methods
         inline void CopyToArray(float* rgfl) const              { rgfl[0] = x, rgfl[1] = y, rgfl[2] = z; }
-        inline float Length(void) const                         { return (float)sqrt(x*x + y*y + z*z); }
+        inline float Length(void) const                         { return std::sqrt(x*x + y*y + z*z); }
         inline operator float *()                               { return &x; } // Vectors will now automatically convert to float * when needed
         inline operator Vector2D () const                       { return (*this).Make2D();      }
 
@@ -106,7 +106,7 @@ public:
 
                 return Vec2;
         }
-        inline float Length2D(void) const                       { return (float)sqrt(x*x + y*y); }
+        inline float Length2D(void) const                       { return std::sqrt(x*x + y*y); }
 
         // Members
         vec_t x, y, z;
